+++ app/ClusterSetup.hs
@@ -11,11 +11,8 @@ module ClusterSetup
   ) where
 
 import           AppConfig                             (RunState (..),
-                                                        authenticate,
                                                         enforcePlaintextAuthenticationPolicy)
 import           Control.Concurrent.STM                (readTVarIO)
-import qualified Control.Monad.State                   as State
-import qualified Data.ByteString                       as BS
 import qualified Data.ByteString.Char8                 as BS8
 import qualified Data.Map.Strict                       as Map
 import           Data.Maybe                            (fromMaybe)
@@ -30,16 +27,12 @@ import           Database.Redis.Cluster.Client         (ClusterAuthentication (.
                                                         ClusterConfig (..),
                                                         createClusterClient,
                                                         createClusterClientWithAuthentication,
-                                                        createClusterClientWithFactories)
-import           Database.Redis.Cluster.ConnectionPool (PoolConfig (PoolConfig))
+                                                        defaultClusterConfig)
 import qualified Database.Redis.Cluster.ConnectionPool as CP
-import           Database.Redis.Command                (ClientState (ClientState),
-                                                        RedisCommands (flushAll))
-import qualified Database.Redis.Command                as RedisCommand
+import           Database.Redis.Command                (RedisCommands (flushAll))
 import           Database.Redis.Connector              (Connector,
                                                         clusterPlaintextConnector,
                                                         clusterTLSConnector)
-import           Database.Redis.Internal.MultiplexPool (createMultiplexPool)
 import           Database.Redis.Resp                   (RespData)
 import           Text.Printf                           (printf)
 
@@ -67,19 +60,7 @@ clusterConfigFromState :: RunState -> ClusterConfig
 clusterConfigFromState state =
   let defaultPort = if useTLS state then 6380 else 6379
       seedNode = NodeAddress (host state) (fromMaybe defaultPort (port state))
-      poolConfig = PoolConfig
-        { CP.maxConnectionsPerNode = 10
-        , CP.connectionTimeout = 300
-        , CP.maxRetries = 3
-        , CP.useTLS = useTLS state
-        }
-  in ClusterConfig
-      { clusterSeedNode = seedNode
-      , clusterPoolConfig = poolConfig
-      , clusterMaxRetries = 3
-      , clusterRetryDelay = 100000
-      , clusterTopologyRefreshInterval = 600
-      }
+  in defaultClusterConfig seedNode
 
 createClusterClientFromStateWithMuxCount :: (Client client) =>
   RunState ->
@@ -87,57 +68,13 @@ createClusterClientFromStateWithMuxCount :: (Client client) =>
   Connector client ->
   IO (ClusterClient client)
 createClusterClientFromStateWithMuxCount state requestedMuxCount connector = do
-  let clusterCfg = clusterConfigFromState state
-      muxCount = max 1 requestedMuxCount
+  let clusterCfg = (clusterConfigFromState state)
+        { clusterMultiplexerCount = requestedMuxCount }
   case clusterAuthentication state of
     Nothing ->
-      if muxCount == 1
-        then createClusterClient clusterCfg connector
-        else createClusterClientWithMuxCount clusterCfg connector muxCount
+      createClusterClient clusterCfg connector
     Just authentication ->
-      if muxCount == 1
-        then createClusterClientWithAuthentication clusterCfg authentication connector
-        else createAuthenticatedClusterClientWithMuxCount
-          clusterCfg authentication connector muxCount
-
-createClusterClientWithMuxCount
-  :: (Client client)
-  => ClusterConfig
-  -> Connector client
-  -> Int
-  -> IO (ClusterClient client)
-createClusterClientWithMuxCount clusterCfg connector muxCount =
-  createClusterClientWithFactories
-    CP.createPool
-    (\boundedConnector _ -> createMultiplexPool boundedConnector muxCount)
-    clusterCfg
-    connector
-
-createAuthenticatedClusterClientWithMuxCount
-  :: (Client client)
-  => ClusterConfig
-  -> ClusterAuthentication
-  -> Connector client
-  -> Int
-  -> IO (ClusterClient client)
-createAuthenticatedClusterClientWithMuxCount clusterCfg authentication connector muxCount =
-  createClusterClientWithFactories
-    CP.createPool
-    (\boundedConnector _ -> createMultiplexPool boundedConnector muxCount)
-    clusterCfg
-    authenticatedConnector
-  where
-    authenticatedConnector addr = do
-      conn <- connector addr
-      let clientState = ClientState conn BS.empty
-      _ <- State.evalStateT (RedisCommand.unRedisCommandClient authenticationAction) clientState
-      return conn
-    authenticationAction =
-      case authentication of
-        ClusterPassword passwordValue ->
-          authenticate "default" (BS8.unpack passwordValue)
-        ClusterACL usernameValue passwordValue ->
-          authenticate (BS8.unpack usernameValue) (BS8.unpack passwordValue)
+      createClusterClientWithAuthentication clusterCfg authentication connector
 
 clusterAuthentication :: RunState -> Maybe ClusterAuthentication
 clusterAuthentication state